Includes new files and work in a more efficient manner(this is under testing)
