+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OpenCV/OpenCV.java
@@ -1,17 +1,13 @@
-package org.firstinspires.ftc.teamcode.Test;
+package org.firstinspires.ftc.teamcode.OpenCV;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfPoint;
-import org.opencv.core.MatOfPoint2f;
 import org.opencv.core.Point;
-import org.opencv.core.Rect;
 import org.opencv.core.Scalar;
 import org.opencv.imgproc.Imgproc;
 import org.openftc.easyopencv.OpenCvCamera;
@@ -33,16 +29,9 @@ import java.util.List;
  * monitor: 640 x 480
  *YES
  */
-@Autonomous(name= "OpenCVmovement")
-public class OpenCVMovement extends LinearOpMode {
+@Autonomous(name= "OpenCVTest")
+public class OpenCV extends LinearOpMode {
     private ElapsedTime runtime = new ElapsedTime();
-
-    //motors
-    private DcMotor leftbackDrive;
-    private DcMotor rightbackDrive;
-    private DcMotor leftfrontDrive;
-    private DcMotor rightfrontDrive;
-
 
     //0 means black, 1 means yellow
     //-1 for debug, but we can keep it like this because if it works, it should change to either 0 or 255
@@ -66,34 +55,14 @@ public class OpenCVMovement extends LinearOpMode {
 
     OpenCvCamera phoneCam;
 
-    public void movement(int left, int right,int time ){
-        leftbackDrive.setPower(left);
-        rightbackDrive.setPower(left);
-        leftfrontDrive.setPower(right);
-        rightfrontDrive.setPower(right);
-        sleep(time);
-    }
-
     @Override
     public void runOpMode() throws InterruptedException {
-        // motors mapping
-        leftbackDrive = hardwareMap.get(DcMotor.class, "BLM");
-        rightbackDrive = hardwareMap.get(DcMotor.class, "BRM");
-        leftfrontDrive = hardwareMap.get(DcMotor.class, "FLM");
-        rightfrontDrive = hardwareMap.get(DcMotor.class, "FRM");
-
-        // Reverse the motor that runs backwards when connected directly to the REV hub
-        leftbackDrive.setDirection(DcMotor.Direction.FORWARD);
-        rightbackDrive.setDirection(DcMotor.Direction.REVERSE);
-        leftfrontDrive.setDirection(DcMotor.Direction.FORWARD);
-        rightfrontDrive.setDirection(DcMotor.Direction.REVERSE);
-
 
         int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
 
         //P.S. if you're using the latest version of easyopencv, you might need to change the next line to the following:
         phoneCam = OpenCvCameraFactory.getInstance().createInternalCamera(OpenCvInternalCamera.CameraDirection.BACK, cameraMonitorViewId);
-         //phoneCam = new OpenCvInternalCamera(OpenCvInternalCamera.CameraDirection.BACK, cameraMonitorViewId);//remove this
+        //phoneCam = new OpenCvInternalCamera(OpenCvInternalCamera.CameraDirection.BACK, cameraMonitorViewId);//remove this
 
         phoneCam.openCameraDevice();//open camera
         phoneCam.setPipeline(new StageSwitchingPipeline());//different stages
@@ -103,7 +72,6 @@ public class OpenCVMovement extends LinearOpMode {
 
         waitForStart();
         runtime.reset();
-
         while (opModeIsActive()) {
             telemetry.addData("Values", valLeft+"   "+valMid+"   "+valRight);
             telemetry.addData("Height", rows);
@@ -111,12 +79,6 @@ public class OpenCVMovement extends LinearOpMode {
 
             telemetry.update();
             sleep(100);
-            if ( time == 5 && valLeft == 0 && valMid == 0 && valRight == 0){
-                movement(1,1,1500);
-                movement(-1,-1,1500);
-            }else if( time == 5 && valLeft == 255 && valMid == 255 && valRight == 255){
-                movement(1,1,1500);
-            }
         }
     }
